+++ src/file.ts
@@ -27,6 +27,13 @@ export interface FileBaseOptions {
    * @default true
    */
   readonly readonly?: boolean;
+
+  /**
+   * Whether the generated file should be marked as executable.
+   *
+   * @default false
+   */
+  readonly executable?: boolean;
 }
 
 export abstract class FileBase extends Component {
@@ -47,6 +54,11 @@ export abstract class FileBase extends Component {
   public readonly: boolean;
 
   /**
+   * Indicates if the file should be marked as executable
+   */
+  public executable: boolean;
+
+  /**
    * The absolute path of this file.
    */
   public readonly absolutePath: string;
@@ -55,6 +67,7 @@ export abstract class FileBase extends Component {
     super(project);
 
     this.readonly = options.readonly ?? true;
+    this.executable = options.executable ?? false;
     this.path = filePath;
 
     this.absolutePath = path.resolve(project.outdir, filePath);
@@ -86,8 +99,9 @@ export abstract class FileBase extends Component {
    * emit.
    * @param resolver Call `resolver.resolve(obj)` on any objects in order to
    * resolve token functions.
+   * @returns the content to synthesize or undefined to skip the file
    */
-  protected abstract synthesizeContent(resolver: IResolver): string;
+  protected abstract synthesizeContent(resolver: IResolver): string | undefined;
 
   /**
    * Writes the file to the project's output directory
@@ -95,9 +109,14 @@ export abstract class FileBase extends Component {
   public synthesize() {
     const outdir = this.project.outdir;
     const filePath = path.join(outdir, this.path);
-    const resolver: IResolver = { resolve: (obj, options) => resolve(obj, [outdir], options) };
-    writeFile(filePath, this.synthesizeContent(resolver), {
+    const resolver: IResolver = { resolve: (obj, options) => resolve(obj, options) };
+    const content = this.synthesizeContent(resolver);
+    if (content === undefined) {
+      return; // skip
+    }
+    writeFile(filePath, content, {
       readonly: this.readonly,
+      executable: this.executable,
     });
   }
 }
@@ -124,4 +143,10 @@ export interface ResolveOptions {
    * @default false
    */
   readonly omitEmpty?: boolean;
+
+  /**
+   * Context arguments.
+   * @default []
+   */
+  readonly args?: any[];
 }