+++ src/typescript.ts
@@ -275,11 +275,13 @@ export class TypeScriptProject extends NodeProject {
     // just run them directly from javascript.
     if (this.jest && compiledTests) {
       this.addDevDeps('@types/jest');
+
       const testout = path.relative(this.srcdir, this.testdir);
       const libtest = path.join(this.libdir, testout);
       const srctest = this.testdir;
 
       this.jest.addTestMatch(`**/${libtest}/**/?(*.)+(spec|test).js?(x)`);
+      this.jest.addWatchIgnorePattern(`/${this.srcdir}/`);
 
       const resolveSnapshotPath = (test: string, ext: string) => {
         const fullpath = test.replace(libtest, srctest);